pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "subway-telegram-bot"
version = "1.0.0"
description = "Production-ready Telegram bot for generating Subway Surfers save-file templates."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "OpenAI Codex" }
]
dependencies = [
  "python-telegram-bot[webhooks]>=22.6,<23",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9",
  "ruff>=0.11,<0.12",
]

[project.scripts]
subway-bot = "subway_bot.main:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["subway_bot*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "B",
  "UP",
]
